+++ src/main/java/fr/maxlego08/essentials/module/ZModuleManager.java
@@ -13,6 +13,7 @@ import fr.maxlego08.essentials.module.modules.chat.ChatModule;
 import fr.maxlego08.essentials.module.modules.customcommands.CustomCommandModule;
 import dev.yanianz.essentials.customscreens.CustomScreenModule;
 import fr.maxlego08.essentials.module.modules.effects.EffectsModule;
+import fr.maxlego08.essentials.module.modules.chatgames.ChatGamesModule;
 import fr.maxlego08.essentials.module.modules.terms.TermsModule;
 import fr.maxlego08.essentials.module.modules.HomeModule;
 import fr.maxlego08.essentials.module.modules.ItemModule;
@@ -89,6 +90,7 @@ public class ZModuleManager implements ModuleManager {
         this.modules.put(CustomScreenModule.class, new CustomScreenModule(this.plugin));
         this.modules.put(EffectsModule.class, new EffectsModule(this.plugin));
         this.modules.put(TermsModule.class, new TermsModule(this.plugin));
+        this.modules.put(ChatGamesModule.class, new ChatGamesModule(this.plugin));
         if (plugin.isPaperVersion()) {
             this.modules.put(DeathMessageModule.class, new DeathMessageModule(this.plugin));
         }